% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{pfa_edd_table}
\alias{pfa_edd_table}
\title{Make a table summarizing the EDD results for all faults.}
\usage{
pfa_edd_table()
}
\description{
This function reproduces the TEP table in the paper.
WARNING: Before running this function
your must either have run \code{run_TEP_sim} yourself, or have downloaded
the results/ and thresholds/ directories from https://github.com/Tveten/tdpcaTEP
and put them in your working directory.
}
\details{
Prints the PFAs and sorts the columns of the output table in the order of PFA.
}
