% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{rl_density}
\alias{rl_density}
\title{Plot density estimates of the run lengths}
\usage{
rl_density(lag, fault_nr, keep_names = NULL,
  change_distr = c("semisparse_uniform", "semisparse_sd_only",
  "semisparse_mean_only"), xlim = c(50, 300))
}
\arguments{
\item{lag}{The lag used in TDPCA.}

\item{fault_nr}{A number between 1 and 20 indicating the fault number.}

\item{keep_names}{A vector of names indicating which methods should be kept for
display (for internal use).}

\item{change_distr}{Which change distributions TDPCA to show results for
(see \code{?tpca::tpca}).}

\item{xlim}{A vector with lower and upper bounds for the x-axis.}
}
\description{
WARNING: Before running this function
your must either have run \code{run_TEP_sim} yourself, or have downloaded
the results/ and thresholds/ directories from https://github.com/Tveten/tdpcaTEP
and put them in your working directory.
}
