% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPCA_functions.R
\name{run_TEP_dpca}
\alias{run_TEP_dpca}
\title{Run the DPCA method on the TEP data}
\usage{
run_TEP_dpca(fault_nr = 1:20, threshold = "analytical")
}
\arguments{
\item{fault_nr}{A numeric vector of desired fault numbers between 1 and 20.}

\item{threshold}{'analytical' or 'val' to determine whether the analytical
thresholds for DPCA should be used, ot thresholds determined by a validation
set.}
}
\description{
This function reproduces the run length results of DPCA on the TEP data.
Results are stored in results/TEP_rl_dpca_<threshold>.txt.
}
