% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TEP_sim.R
\name{run_TEP_sim}
\alias{run_TEP_sim}
\title{Reproduce the TEP simulation study.}
\usage{
run_TEP_sim()
}
\description{
This function reproduces the results in
"Online Detection of Sparse Changes in High-Dimensional Data Streams Using Tailored Projections"
for the TEP data (Section 5). On an average computer it takes about two days to run on three cores.
Thus, the files with the results of running \code{run_TEP_sim} used in the paper
are included in the package or found at github.com/Tveten/tdpcaTEP.
}
\details{
The TEP data are automatically loaded from the github repository to the
global environement.

The following directories are created in the current directory:
./thresholds,
./thresholds/axes,
.thresholds/threshold_results,
./results,
./results/figures.
Files created to save information about thresholds are saved in the threshold
directory, while EDD results are stored in "results".
The results we obtained by running \code{run_TEP_sim} are stored in the
same structure of directories on Github.
}
