% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arl_functions.R
\name{tdpca_arl}
\alias{tdpca_arl}
\title{Estimate average run length (ARL) for monitoring by tdpca}
\usage{
tdpca_arl(threshold, x_train, lag, axes, n, w, n_sim)
}
\arguments{
\item{threshold}{A numeric specifying the threshold value for when a change
is declared.}

\item{x_train}{The training data (not lag-extended).}

\item{lag}{The number of lags used in TDPCA.}

\item{axes}{A vector specifying which axes/projections to use.}

\item{n}{The number of observations to monitor for an estimate of the ARL.
See details.}

\item{w}{The window size. Number of recent time-points to consider for a
change.}

\item{n_sim}{The number of simulations to base the estimate on.}
}
\value{
An estimate of the average run length.
}
\description{
Estimate average run length (ARL) for monitoring by tdpca
}
