% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{TEP_summary}
\alias{TEP_summary}
\title{Make a table of EDD results and PFA for all tested methods.}
\usage{
TEP_summary(lag, fault_nr, keep_names = NULL,
  change_distr = c("semisparse_uniform", "semisparse_sd_only",
  "semisparse_mean_only"), sort_by = NULL, show_plot = TRUE, ...)
}
\arguments{
\item{lag}{The lag used in TDPCA.}

\item{fault_nr}{A number between 1 and 20 indicating the fault number.}

\item{keep_names}{A vector of names indicating which methods should be kept for
display (for internal use).}

\item{change_distr}{Which change distributions TDPCA to show results for
(see \code{?tpca::tpca}).}

\item{sort_by}{If not NULL, either 'edd' or 'pfa', to sort the output table.}

\item{show_plot}{Logical. Should a density plot also be shown?}
}
\description{
WARNING: Before running this function
your must either have run \code{run_TEP_sim} yourself, or have downloaded
the results/ and thresholds/ directories from https://github.com/Tveten/tdpcaTEP
and put them in your working directory.
}
