% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_finder.R
\name{threshold_finder_tdpca}
\alias{threshold_finder_tdpca}
\title{Estimating the threshold for tpca changepoint detection}
\usage{
threshold_finder_tdpca(x_train, lag, n, alpha, axes = NULL, w = 200,
  rel_tol = c(0.2, 0.1, 0.05, 0.025), thresh_alpha = 0.05,
  init_thresh = NULL, learning_coef = NULL, file_id = NULL)
}
\arguments{
\item{x_train}{The d x m training data matrix, where d is the dimension of the data
and m the number of training samples.}

\item{lag}{The number of lags in TDPCA and the block bootstrap.}

\item{n}{The length of the segment to monitor for false alarms. See details.}

\item{alpha}{Probability of type I error (false alarm) within the time window . See details.}

\item{axes}{Indices of the principal axes to be used in simulations.
MUST be supplied if mon_type == 'tpca'.}

\item{w}{The window size (an integer).}

\item{rel_tol}{A vector with the sequence of relative error tolerances
allowed at each step towards convergence in the algorithm. See details.}

\item{thresh_alpha}{1 - \code{thresh_alpha} is the confidence level. See details.}

\item{init_thresh}{Use if a custom initial value for the threshold is wanted.}

\item{learning_coef}{The learning rate is defined as \code{rel_tol} /
\code{learning_coef}. The default \code{learning_coef} is 3, which has
been found a good choice after a lot of experimenting.
#' @param file_id A string that identifies the files associated with the particular
threshold beyond the data dimension d, axes/p0, training size m or
probability of false alarm alpha.}

\item{p0}{The mixture probability of a dimension being affected by a change.
MUST be supplied if mon_type == 'mixture'.}
}
\value{
A list with the following components:
\describe{
  \item{\code{threshold}}{The final estimate }
}
\code{tpca_threshold} also creates a .txt file with each line showing
summaries of each step in the estimation procedure. The last line corresponds
to the final estimate.
}
\description{
Description
}
\details{
\code{n} and \code{alpha} governs the false alarm rate by the relation
  P(T < n | H_0) <= \code{alpha}.
The corresponding average run length (\code{arl}) is approximately given by n / alpha.

\code{rel_tol} and \code{thresh_alpha} governs the number of simulations used in
each step of the algorithm towards a more and more certain estimate.
At each step, the number of simulations is chosen so that
\code{[(1 - rel_tol)arl, (1 + rel_tol)arl]} approximately covers the true
average run length at confidence level \code{thresh_alpha}.
For example, when the last \code{rel_tol} is 0.025, it means that the final
estimated threshold corresponds to an average run length of approximately
\code{arl} +- 0.025 * \code{arl} at confidence level \code{thresh_alpha}.
The algorithm should start with a large relative error tolerance,
and then narrow it down for the quickest convergence.
}
