% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TEP_sim.R
\name{load_TEP_data_globally}
\alias{load_TEP_data_globally}
\title{Load the TEP data to the global environment.}
\usage{
load_TEP_data_globally(no_test = FALSE)
}
\description{
This function loads the TEP data from
https://github.com/Tveten/tdpcaTEP/tree/master/TEP_data/,
and puts it in the global environment.
}
